+++ lib/src/widgets/animations/rive/emoji_enum.dart
@@ -21,8 +21,7 @@ enum RiveEmoji {
   laughing('Laughing'),
 
   ///
-  surprise('Surprise')
-  ;
+  surprise('Surprise');
 
   const RiveEmoji(this.artBoard);